Removed old test code and updated sweep_parameter function (see description for more)
The sweep_parameter function was updated to snap values instead of sample over a period of time. This is faster. Sampling variables over a period of time didn't seem to offer any accuracy advantages (as the time constant of the lock-in should do averaging by itself). Variable names were updated and the wrapper calls were updated to reflect the fact that the pasternack rf source is now being used instead of the HP8350B.